rel 8; try a second send if there's a resource name conflict
If sending fails due to stream error, most likely due to a resource name
conflict, wait a second and try to send again with a randomized resource
name.
For jabber clients with <thread> support these two past changes won't be
noticeable, however for the rest they'll mean that 99% of the times
msgs come from a constant jid/resource pair with only a very occasional
msg coming from a randomized resource.